tareas 5 y 6

+++ senderos/views.py
@@ -1,11 +1,21 @@
 from django.shortcuts import render, redirect
 from mongoengine.queryset.visitor import Q
-from .models import Excursion, Fotos
-from .forms import ExcursionForm
+from .models import Excursion, Fotos, ExcursionSerializer
+from .forms import ExcursionForm, RegisterForm
 from django.contrib import messages
 from django.conf import settings
 import os
 import shutil
+from django.contrib.auth import authenticate, login
+from django.contrib.auth.signals import user_logged_out
+import logging
+from rest_framework.views import APIView
+from rest_framework.response import Response
+from django.http import Http404
+from rest_framework import status
+from rest_framework.permissions import IsAuthenticated
+
+logger = logging.getLogger(__name__)
 
 IMAGE_DIR = os.path.join(settings.BASE_DIR, 'senderos', 'static', 'img', 'senderos')
 
@@ -18,6 +28,8 @@ def index(request):
 		'form': form
 	}
 
+	logger.info('Página principal')
+
 	return render(request, 'senderos/index.html', context)
 
 def buscar(request):
@@ -27,6 +39,8 @@ def buscar(request):
 		'excursiones': Excursion.objects(Q(nombre__icontains=busqueda) | Q(descripcion__icontains=busqueda))
 	}
 	
+	logger.info('Buscador de rutas')
+
 	return render(request, 'senderos/index.html', context)
 
 def editar(request, id):
@@ -46,8 +60,10 @@ def editar(request, id):
 				guardarImagenes(id, files, e, input_d)
 
 			messages.success(request, "Ruta editada con éxito")
+			logger.info('Ruta editada correctamente')
 
 		else:
+			logger.exception('Error en el formulario al editar una ruta')
 			messages.error(request, "Error en el formulario")
 
 	return redirect('index')
@@ -66,7 +82,10 @@ def aniadir(request):
 
 			messages.success(request, "Ruta añadida con éxito")
 
+			logger.info('Ruta añadida correctamente')
+
 		else:
+			logger.exception('Error en el formulario de añadir una ruta')
 			messages.error(request, "Error en el formulario")
 
 	return redirect('index')
@@ -89,8 +108,10 @@ def guardarImagenes(id, files, e, input_d):
 			e.fotos.append(Fotos(pie=input_d.get('pie'), file=input_d.get('foto').name))
 
 		e.save()
+		logger.info('Imágenes de ruta subidas correctamente')
 
 	except OSError as error:
+		logger.exception('Error al subir foto en una ruta')
 		print('**************** ERROR', error)
 
 def info(request, id):
@@ -106,13 +127,88 @@ def info(request, id):
 		'form': form
 	}
 	
+	logger.info('Vista de una ruta')
+
 	return render(request, 'senderos/info.html', context)
 
 def eliminar(request, id):
 	directorio = os.path.join(IMAGE_DIR, id)
 
-	shutil.rmtree(directorio)
+	if os.path.exists(directorio):
+		shutil.rmtree(directorio)
 
 	Excursion.objects(id=id).delete()
 	
+	logger.info('Eliminación de una ruta')
+
 	return redirect('index')
+
+def registrar(request):
+	if request.method == "POST":
+		form = RegisterForm(request.POST)
+		if form.is_valid():
+			form.save()
+			new_user = authenticate(username=form.cleaned_data['username'],
+															password=form.cleaned_data['password1'],
+															)
+			login(request, new_user)
+			logger.info('Se ha registrado y logeado un usuario')
+
+		else:
+			logger.exception('Error en el formulario al registrar usuario')
+
+		return redirect("/")
+
+	else:
+		logger.exception('Error en el POST de registrar usuario')
+		form = RegisterForm()
+
+	return render(request, "registration/register.html", {"form":form})
+
+class ExcursionesView(APIView):
+
+	permission_classes = (IsAuthenticated,)
+
+	def get(self, request):
+		excursiones = Excursion.objects.all()[:4]
+		serializer = ExcursionSerializer(excursiones, many=True)
+		return Response(serializer.data)
+
+	def post(self, request):
+		serializer = ExcursionSerializer(data=request.data)
+
+		if serializer.is_valid():
+			serializer.save()
+			return Response(serializer.data, status=status.HTTP_201_CREATED)
+		return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
+class ExcursionView(APIView):
+
+	def get(self, request, id):
+		try:
+			e = Excursion.objects.get(id=id)
+			serializer = ExcursionSerializer(e)
+			return Response(serializer.data)
+		except:
+			raise Http404
+
+	def delete(self, request, id):
+		try:
+			e = Excursion.objects.get(id=id)
+			e.delete()
+			serializer = ExcursionSerializer(e)
+			return Response(status=status.HTTP_204_NO_CONTENT)
+		except:
+			raise Http404
+
+	def put(self, request, id):
+		try:
+			e = Excursion.objects.get(id=id)
+			serializer = ExcursionSerializer(e, data=request.data)
+			if serializer.is_valid():
+				serializer.save()
+				return Response(serializer.data)
+			return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
+		except:
+			raise Http404
